backend/requirements.txt
# FastAPI Backend Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0
redis==5.0.1

# Deep Learning (reuse from main requirements)
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.3.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# Utilities
requests>=2.31.0
aiofiles==23.2.1
